moved launch tasks to individual classes. made ~/.tracerrc the default config file
