+++ tracer.conf
@@ -2,15 +2,15 @@
 
 [tool_locations]
 #paths to tools used by TraCeR for alignment, quantitation, etc
-bowtie2_path = /path/to/bowtie2
-igblast_path = /path/to/igblastn
-kallisto_path = /path/to/kallisto
-trinity_path = /path/to/trinity
-dot_path = /path/to/dot
-neato_path = /path/to/neato
+#bowtie2_path = /path/to/bowtie2
+#igblast_path = /path/to/igblastn
+#kallisto_path = /path/to/kallisto
+#trinity_path = /path/to/trinity
+#dot_path = /path/to/dot
+#neato_path = /path/to/neato
 
 [bowtie2_options]
-synthetic_genome_index_path = resources/mouse/synthetic_genomes
+synthetic_genome_index_path = resources/Mmus/synthetic_genomes
 
 [trinity_options]
 #line below specifies maximum memory for Trinity Jellyfish component. Set it appropriately for your environment.
@@ -25,8 +25,8 @@ max_jellyfish_memory = 1G
 
 
 [IgBlast_options]
-igblast_index_location = resources/mouse/igblast_dbs
-imgt_seq_location = resources/mouse/imgt_sequences
+#igblast_index_location = resources/Mmus/igblast_dbs
+#imgt_seq_location = resources/Mmus/imgt_sequences
 igblast_seqtype = TCR
 
 [kallisto_options]
