+++ tracerlib/__init__.py
@@ -2,4 +2,3 @@ import os
 
 # Parent directory
 base_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
-resource_lookup = {'Mmus': 'mouse', 'HSap': 'human'}
